refactor: move ensure_secret_key_exists into key.rs
